+++ src/compact/index.ts
@@ -1,7 +1,8 @@
 /**
- * compacts a given array. It removes all falsey values EXCEPT the number 0.
+ * compacts a given array. It removes all falsey values EXCEPT the number 0 by default.
+ * You can also filter the 0 values by setting the `withZero` arg to true.
  */
-const compact = (array: Array<any> = [], withZero?: boolean = false): Array<any> => {
+const compact = (array: any[] = [], withZero?: boolean): any[] => {
   try {
     return array.filter((item: number | string) => {
       if (withZero) {
@@ -11,7 +12,7 @@ const compact = (array: Array<any> = [], withZero?: boolean = false): Array<any>
     });
   } catch (e) {
     if (!Array.isArray(array)) {
-      throw `First argument of js-extra.compact() should be an array. ${array} given.`;
+      throw new Error(`First argument of js-extra.compact() should be an array. ${array} given. ${e}`);
     }
     throw e;
   }